Website for the Spring Semester 2020
including new teachers and old content. In order to make it a bit easier
managing the content and possibly copying exercises or other content around,
I added a Previous Semesters section, including all content from last two semesters.
In order to keep the Author tag working and not flood the home page with a list of
teachers that would be too long, I divided the teachers list in 'current' and 'previous'.
Only current teachers are shown on the home page, while everybody used when getting
the data to show name and avatar.

I also bumped up the version of vuepress and had to therefore slightly change the
way the sidebar is drawn. Since I was at it, I made it automatic for new files to
be picked up, so there's no need to edit configuration after having added a new lesson.